removed warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5001 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/od2trips/ODDistrict.cpp
@@ -90,14 +90,14 @@ ODDistrict::getRandomSink() const throw(OutOfBoundsException)
 unsigned int
 ODDistrict::sinkNumber() const
 {
-    return mySinks.getVals().size();
+    return (unsigned int) mySinks.getVals().size();
 }
 
 
 unsigned int
 ODDistrict::sourceNumber() const
 {
-    return mySources.getVals().size();
+    return (unsigned int) mySources.getVals().size();
 }
